refactor(agents): move swarm family into subagents/swarm
