+++ src/agents/openclaw-tools.swarm.ts
@@ -4,7 +4,7 @@ import {
   getSubagentRunByRunId,
   recordSwarmStructuredOutput,
 } from "./subagent-registry.js";
-import { resolveSwarmConfig } from "./swarm-config.js";
+import { resolveSwarmConfig } from "./subagents/swarm/swarm-config.js";
 import { createAgentsWaitTool } from "./tools/agents-wait-tool.js";
 import type { AnyAgentTool } from "./tools/common.js";
 import { createStructuredOutputTool } from "./tools/structured-output-tool.js";
