+++ src/agents/subagent-registry-lifecycle-completion.ts
@@ -33,8 +33,8 @@ import type {
 import type { createSubagentRegistryLifecycleDelivery } from "./subagent-registry-lifecycle-delivery.js";
 import { createSubagentRegistryLifecycleTerminalCleanup } from "./subagent-registry-lifecycle-terminal-cleanup.js";
 import type { SubagentCompletionRequest, SubagentRunRecord } from "./subagent-registry.types.js";
-import { updateSwarmCollectorCompletion } from "./swarm-collector.js";
-import { releaseSwarmRun } from "./swarm-scheduler.js";
+import { updateSwarmCollectorCompletion } from "./subagents/swarm/swarm-collector.js";
+import { releaseSwarmRun } from "./subagents/swarm/swarm-scheduler.js";
 import { peekSwarmStructuredOutput } from "./tools/structured-output-tool.js";
 
 export function createSubagentRegistryLifecycleCompletion(
