+++ src/agents/subagent-registry-restore.ts
@@ -23,8 +23,8 @@ import {
 } from "./subagent-session-reconciliation.js";
 import { retrySubagentCleanup } from "./subagent-spawn-cleanup.js";
 import { readGatewayRunId } from "./subagent-spawn-gateway.js";
-import { resolveSwarmConfig } from "./swarm-config.js";
-import { enqueueSwarmRun } from "./swarm-scheduler.js";
+import { resolveSwarmConfig } from "./subagents/swarm/swarm-config.js";
+import { enqueueSwarmRun } from "./subagents/swarm/swarm-scheduler.js";
 
 type RestoredQueuedFailureSettlementClaim = {
   entry: SubagentRunRecord;
