+++ src/agents/subagents/swarm/swarm-config.ts
@@ -1,7 +1,7 @@
 import { isRecord } from "@openclaw/normalization-core/record-coerce";
-import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { clampNumber } from "../utils.js";
-import { resolveAgentConfig } from "./agent-scope-config.js";
+import type { OpenClawConfig } from "../../../config/types.openclaw.js";
+import { clampNumber } from "../../../utils.js";
+import { resolveAgentConfig } from "../../agent-scope-config.js";
 
 type ResolvedSwarmConfig = {
   enabled: boolean;
