+++ src/agents/subagents/swarm/swarm-scheduler.test-support.ts
