+++ src/agents/tools/agents-wait-tool.ts
@@ -5,7 +5,7 @@ import { resolveSubagentCompletionResultText } from "../subagent-completion-resu
 import { onSubagentRegistryPersisted } from "../subagent-registry-state.js";
 import { getSubagentRunsByRunIds } from "../subagent-registry.js";
 import type { SubagentRunRecord } from "../subagent-registry.types.js";
-import { resolveSwarmConfig } from "../swarm-config.js";
+import { resolveSwarmConfig } from "../subagents/swarm/swarm-config.js";
 import type { AnyAgentTool } from "./common.js";
 import { jsonResult, ToolInputError } from "./common.js";
