+++ src/gateway/agent-turn/agent-request-preflight.ts
@@ -8,8 +8,8 @@ import {
   findAuthorizedSwarmCollectorRequest,
   findSwarmCollectorSession,
 } from "../../agents/subagent-registry-memory.js";
-import { resolveSwarmConfig } from "../../agents/swarm-config.js";
-import { validateStructuredOutputSchema } from "../../agents/swarm-output-schema.js";
+import { resolveSwarmConfig } from "../../agents/subagents/swarm/swarm-config.js";
+import { validateStructuredOutputSchema } from "../../agents/subagents/swarm/swarm-output-schema.js";
 import { resolveAgentIdFromSessionKey, resolveStorePath } from "../../config/sessions.js";
 import { loadSessionEntry } from "../../config/sessions/session-accessor.js";
 import {
